Get prices with the prices hook and change label formatting
