+++ src/components/one-time-donation/components/ThirdStep.tsx
@@ -1,6 +1,6 @@
 import { Collapse, Grid, List, Divider, Typography } from '@mui/material'
 import { createStyles, makeStyles } from '@mui/styles'
-import RadioGroupFormik from './RadioGroupFormik'
+import PriceRadioGroup from './PriceRadioGroup'
 import { useField } from 'formik'
 import { UseQueryResult } from 'react-query'
 import { useRouter } from 'next/router'
@@ -75,7 +75,7 @@ export default function ThirdStep() {
         <Typography className={classes.h2}>{t('third-step.title')}</Typography>
       </Grid>
       <Grid item mx={10}>
-        <RadioGroupFormik name="payment" options={options as []} />
+        <PriceRadioGroup name="payment" options={options as []} />
       </Grid>
       <Grid>
         <Collapse in={field.value === 'bank'} timeout="auto" unmountOnExit>
